remove instrument manager

+++ algotrader-core/src/main/java/com/unisoft/algotrader/event/LogMarketDataEventBus.java
@@ -15,20 +15,20 @@ public class LogMarketDataEventBus implements EventBus.MarketDataEventBus {
     }
 
     @Override
-    public void publishBar(int instId, int size, long dateTime, double open, double high, double low, double close, long volume, long openInt) {
+    public void publishBar(long instId, int size, long dateTime, double open, double high, double low, double close, long volume, long openInt) {
         LOG.info("publishBar, altInstId = {}, size = {}, dateTime = {}, open = {}, high = {}, low = {}, close = {}, volume = {}, openInt = {}"
                 , instId, size, dateTime, open, high, low, close, volume, openInt);
 
     }
 
     @Override
-    public void publishQuote(int instId, long dateTime, double bid, double ask, int bidSize, int askSize) {
+    public void publishQuote(long instId, long dateTime, double bid, double ask, int bidSize, int askSize) {
         LOG.info("publishQuote, altInstId = {}, dateTime = {}, bid = {}, ask = {}, bidSize = {}, askSize = {}"
                 , instId, dateTime, bid, ask, bidSize, askSize);
     }
 
     @Override
-    public void publishTrade(int instId, long dateTime, double price, int size) {
+    public void publishTrade(long instId, long dateTime, double price, int size) {
         LOG.info("publishTrade, altInstId = {}, dateTime = {}, price = {}, size = {}"
                 , instId, dateTime, price, size);
     }
