+++ algotrader-core/src/main/java/com/unisoft/algotrader/event/RingBufferMarketDataEventBus.java
@@ -16,7 +16,7 @@ public class RingBufferMarketDataEventBus implements EventBus.MarketDataEventBus
     }
 
     @Override
-    public void publishBar(int instId, int size, long dateTime, double open, double high, double low, double close, long volume, long openInt) {
+    public void publishBar(long instId, int size, long dateTime, double open, double high, double low, double close, long volume, long openInt) {
 
         long sequence = marketDataRB.next();
         MarketDataContainer event = marketDataRB.get(sequence);
@@ -32,7 +32,7 @@ public class RingBufferMarketDataEventBus implements EventBus.MarketDataEventBus
     }
 
     @Override
-    public void publishQuote(int instId, long dateTime, double bid, double ask, int bidSize, int askSize) {
+    public void publishQuote(long instId, long dateTime, double bid, double ask, int bidSize, int askSize) {
         long sequence = marketDataRB.next();
         MarketDataContainer event = marketDataRB.get(sequence);
         event.reset();
@@ -46,7 +46,7 @@ public class RingBufferMarketDataEventBus implements EventBus.MarketDataEventBus
     }
 
     @Override
-    public void publishTrade(int instId, long dateTime, double price, int size) {
+    public void publishTrade(long instId, long dateTime, double price, int size) {
         long sequence = marketDataRB.next();
         MarketDataContainer event = marketDataRB.get(sequence);
         event.reset();
