+++ algotrader-persistence/algotrader-persistence-cassandra/src/main/java/com/unisoft/algotrader/persistence/cassandra/accessor/ExecutionReportAccessor.java
@@ -23,7 +23,7 @@ public interface ExecutionReportAccessor {
     ListenableFuture<Result<ExecutionReport>> getAllAsync();
 
     @Query("SELECT * FROM trading.execution_reports WHERE inst_id = :inst_id")
-    Result<ExecutionReport> getByInstId(@Param("inst_id") int inst_id);
+    Result<ExecutionReport> getByInstId(@Param("inst_id") long inst_id);
 
     @Query("SELECT * FROM trading.execution_reports WHERE order_id = :order_id")
     Result<ExecutionReport> getByOrderId(@Param("order_id") long order_id);
